+++ RELEASES.md
@@ -58,6 +58,7 @@ Upcoming
 - Add Test.properties to AppConstants for local properties which should not be added to the version control system
 - Bugfix ForEachChildElementPipe with blockSize which was skipping remaining items after last block
 - Remove datasource.deleteTable function from Larva
+- Remove autoignore function from Larva